+++ backend/store/setting.go
@@ -5,6 +5,7 @@ import (
 	"database/sql"
 	"fmt"
 	"strings"
+	"sync"
 
 	"github.com/pkg/errors"
 	"google.golang.org/protobuf/encoding/protojson"
@@ -94,8 +95,8 @@ func (s *Store) PatchSetting(ctx context.Context, patch *api.SettingPatch) (*api
 	return setting.toAPISetting(), nil
 }
 
-func (s *Store) DeleteSettingCache(name api.SettingName) {
-	s.settingCache.Delete(name)
+func (s *Store) DeleteCache() {
+	s.settingCache = sync.Map{}
 }
 
 // FindSettingMessage is the message for finding setting.
